requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.32
python-decouple==3.8
pyngrok==6.0.0
nltk==3.8.1
textblob==0.17.1
tweepy==4.14.0